Restructure the lib, make HD derivations optional
